Feature/search
## やったこと
- Model の `get_reference_items_by_keyword`を作成
  - select('reference_items.*')
  - distinct()
- Model の `search_tasks_by_keyword`を作成
- controller で `action_search`をajaxで実装
- knockout.jsでリアルタイム検索
## 動作確認
- tasks, referenceitems のリアルタイム検索
- tasks は `title`, `description` から検索
- referenceitems は `title`, `tags` から検索

+++ fuel/app/classes/model/task.php
@@ -2,16 +2,17 @@
 
 class Model_Task extends \Model
 {
-    public static function get_tasks_all()
+    public static function get_tasks_all($user_id)
     {
         return \DB::select()
             ->from('tasks')
+            ->where('user_id', '=', $user_id)
             ->order_by('created_at', 'desc')
             ->execute()
             ->as_array();
     }
 
-    public static function get_task_by_id($task_id)
+    public static function get_task_by_id($task_id, $user_id)
     {
         return \DB::select(
             'tasks.*',
@@ -20,13 +21,20 @@ class Model_Task extends \Model
             ->from('tasks')
             ->join('users')
             ->on('tasks.user_id', '=', 'users.id')
-            ->where('tasks.id', $task_id)
+            ->where('tasks.id', '=', $task_id)
+            ->where('tasks.user_id', '=', $user_id)
             ->execute()
             ->current();
     }
 
-    public static function create_task($title, $description, $status, $dev_location, $deadline, $user_id)
-    {
+    public static function create_task(
+        $title,
+        $description,
+        $status,
+        $dev_location,
+        $deadline,
+        $user_id
+    ) {
         return \DB::insert('tasks')
             ->set(array(
                 'title' => $title,
@@ -41,9 +49,15 @@ class Model_Task extends \Model
             ->execute();
     }
 
-
-    public static function update_task($task_id, $title, $description, $status, $dev_location, $deadline)
-    {
+    public static function update_task(
+        $task_id,
+        $title,
+        $description,
+        $status,
+        $dev_location,
+        $deadline,
+        $user_id
+    ) {
         return \DB::update('tasks')
             ->set(array(
                 'title' => $title,
@@ -53,16 +67,30 @@ class Model_Task extends \Model
                 'deadline' => $deadline,
                 'updated_at' => date('Y-m-d H:i:s'),
             ))
-            ->where('id', $task_id)
+            ->where('id', '=', $task_id)
+            ->where('user_id', '=', $user_id)
             ->execute();
     }
 
-
-    public static function delete_task($task_id)
+    public static function delete_task($task_id, $user_id)
     {
         return \DB::delete('tasks')
-            ->where('id', $task_id)
+            ->where('id', '=', $task_id)
+            ->where('user_id', '=', $user_id)
             ->execute();
     }
 
+    public static function search_tasks_by_keyword($user_id, $keyword)
+    {
+        return \DB::select()
+            ->from('tasks')
+            ->where('user_id', '=', $user_id)
+            ->where_open()
+                ->where('title', 'like', '%' . $keyword . '%')
+                ->or_where('description', 'like', '%' . $keyword . '%')
+            ->where_close()
+            ->order_by('created_at', 'desc')
+            ->execute()
+            ->as_array();
+    }
 }
